# openapi-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/schedule": {
      "post": {
        "operationId": "ScheduleController_createSchedule",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkCreateMentoringScheduleDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "정기 스케줄이 성공적으로 등록되었습니다."
          },
          "403": {
            "description": "본인의 스케줄만 등록할 수 있습니다."
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "멘토 정기 스케줄 등록",
        "tags": [
          "Schedule"
        ]
      },
      "patch": {
        "operationId": "ScheduleController_updateSchedule",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkUpdateMentoringScheduleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "정기 스케줄이 성공적으로 수정되었습니다."
          },
          "403": {
            "description": "본인의 스케줄만 수정할 수 있습니다."
          },
          "404": {
            "description": "해당 스케줄을 찾을 수 없습니다."
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "멘토 정기 스케줄 수정",
        "tags": [
          "Schedule"
        ]
      },
      "get": {
        "operationId": "ScheduleController_getMyScheduleList",
        "parameters": [],
        "responses": {
          "200": {
            "description": "멘토가 등록한 정기 스케줄 목록을 반환합니다.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetScheduleListResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "멘토 정보를 찾을 수 없습니다."
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "멘토 정기 스케줄 목록 조회",
        "tags": [
          "Schedule"
        ]
      }
    },
    "/schedule/{id}": {
      "delete": {
        "operationId": "ScheduleController_deleteSchedule",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "정기 스케줄이 성공적으로 삭제되었습니다."
          },
          "403": {
            "description": "본인의 스케줄만 삭제할 수 있습니다."
          },
          "404": {
            "description": "해당 스케줄을 찾을 수 없습니다."
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "멘토 정기 스케줄 삭제",
        "tags": [
          "Schedule"
        ]
      }
    },
    "/schedule/reservations": {
      "get": {
        "operationId": "ScheduleController_getMentorReservationList",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "페이지 번호",
            "schema": {
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "페이지당 항목 수",
            "schema": {
              "default": 10,
              "example": 10,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "등록된 세션 목록",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MentorReservationListResponseDto"
                  }
                }
              }
            }
          },
          "500": {
            "description": "등록된 세션 목록을 찾을 수 없습니다."
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "멘토 예약 목록 조회",
        "tags": [
          "Schedule"
        ]
      }
    },
    "/schedule/reservations/{id}": {
      "get": {
        "operationId": "ScheduleController_getMentorReservationDetail",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "예약 상세 정보 반환",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MentorReservationDetailResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "멘토 예약 상세 조회",
        "tags": [
          "Schedule"
        ]
      }
    },
    "/schedule/reservations/{id}/status": {
      "patch": {
        "operationId": "ScheduleController_updateReservationStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "description": "예약 상태 변경 요청",
          "content": {
            "application/json": {
              "schema": {
                "type": "string"
              },
              "examples": {
                "수락": {
                  "summary": "예약 수락",
                  "value": {
                    "status": "confirmed"
                  }
                },
                "거절": {
                  "summary": "예약 거절 + 환불 처리",
                  "value": {
                    "status": "rejected",
                    "rejectReason": "일정이 맞지 않아 진행이 어렵습니다."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "예약 상태 변경 성공"
          },
          "400": {
            "description": "요청 형식 또는 조건 오류"
          },
          "403": {
            "description": "권한 없음"
          },
          "404": {
            "description": "예약 없음"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "멘토 예약 수락/거절 처리",
        "tags": [
          "Schedule"
        ]
      }
    },
...
    "/review": {
      "post": {
        "operationId": "ReviewController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateReviewDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "후기를 작성하셨습니다."
          },
          "400": {
            "description": "잘못된 요청. 유효하지 않은 예약 ID이거나 예약 상태가 완료되지 않음."
          },
          "404": {
            "description": "예약 또는 유저를 찾을 수 없음."
          },
          "409": {
            "description": "이미 후기를 작성한 예약입니다."
          },
          "500": {
            "description": "후기 작성 중 서버 오류가 발생했습니다."
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "멘토링 후기 작성",
        "tags": [
          "Review"
        ]
      }
    },
    "/review/{id}": {
      "patch": {
        "operationId": "ReviewController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateReviewDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "후기 수정 성공"
          },
          "403": {
            "description": "작성자 본인만 수정 가능"
          },
          "404": {
            "description": "리뷰를 찾을 수 없음"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "후기 수정",
        "tags": [
          "Review"
        ]
      },
      "delete": {
        "operationId": "ReviewController_Delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "후기 삭제 성공"
          },
          "403": {
            "description": "작성자 본인만 삭제 가능"
          },
          "404": {
            "description": "리뷰를 찾을 수 없음"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "후기 삭제",
        "tags": [
          "Review"
        ]
      }
    },
    "/review/my": {
      "get": {
        "operationId": "ReviewController_getMyReviews",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "페이지 번호 (기본값: 1)",
            "schema": {
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "페이지당 개수 (기본값: 10)",
            "schema": {
              "default": 10,
              "example": 10,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "내가 쓴 후기 조회",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReviewMyListItemDto"
                  }
                }
              }
            }
          },
          "404": {
            "description": "해당 세션을 찾을 수 없습니다."
          },
          "500": {
            "description": "서버 오류"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "내가 쓴 후기 조회",
        "tags": [
          "Review"
        ]
      }
    },
    "/review/received": {
      "get": {
        "operationId": "ReviewController_getMentorReceivedReviews",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "페이지 번호 (기본값: 1)",
            "schema": {
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "페이지당 개수 (기본값: 10)",
            "schema": {
              "default": 10,
              "example": 10,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "내가 받은 후기 조회 (멘토)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReviewReceivedListItemDto"
                  }
                }
              }
            }
          },
          "404": {
            "description": "해당 세션을 찾을 수 없습니다."
          },
          "500": {
            "description": "서버 오류"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "내가 받은 후기 조회 (멘토)",
        "tags": [
          "Review"
        ]
      }
    },
...
    "/session/{id}/reviews": {
      "get": {
        "operationId": "SessionController_getReviews",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "페이지 번호 (기본값: 1)",
            "schema": {
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "페이지당 개수 (기본값: 10)",
            "schema": {
              "default": 10,
              "example": 10,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "세션에 등록된 후기 목록",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReviewListItemDto"
                  }
                }
              }
            }
          },
          "404": {
            "description": "해당 세션을 찾을 수 없습니다."
          },
          "500": {
            "description": "서버 오류"
          }
        },
        "summary": "세션별 후기 조회",
        "tags": [
          "Session"
        ]
      }
    },
...
  },
  "info": {
    "title": "커넥트 api문서",
    "description": "커넥트 개발을 위한 api문서 \n\n🔗 OpenAPI 명세 다운로드: /openapi-spec.json`,",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "JWT Bearer Token"
      }
    },
    "schemas": {
...
      "CreateMentoringScheduleDto": {
        "type": "object",
        "properties": {
          "dayOfWeek": {
            "type": "string",
            "enum": [
              "MONDAY",
              "TUESDAY",
              "WEDNESDAY",
              "THURSDAY",
              "FRIDAY",
              "SATURDAY",
              "SUNDAY"
            ]
          },
          "startTime": {
            "type": "string",
            "example": "10:00"
          },
          "endTime": {
            "type": "string",
            "example": "12:00"
          }
        },
        "required": [
          "dayOfWeek",
          "startTime",
          "endTime"
        ]
      },
      "BulkCreateMentoringScheduleDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateMentoringScheduleDto"
            }
          }
        },
        "required": [
          "data"
        ]
      },
      "UpdateMentoringScheduleDto": {
        "type": "object",
        "properties": {
          "dayOfWeek": {
            "type": "string",
            "enum": [
              "MONDAY",
              "TUESDAY",
              "WEDNESDAY",
              "THURSDAY",
              "FRIDAY",
              "SATURDAY",
              "SUNDAY"
            ]
          },
          "startTime": {
            "type": "string"
          },
          "endTime": {
            "type": "string"
          },
          "id": {
            "type": "string",
            "description": "정기스케줄 ID",
            "example": "b2a9c1ef-27f4-4c02-ae34-0fcecc30c9f1"
          }
        },
        "required": [
          "id"
        ]
      },
      "BulkUpdateMentoringScheduleDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UpdateMentoringScheduleDto"
            }
          }
        },
        "required": [
          "data"
        ]
      },
      "ScheduleItemResponseDto": {
        "type": "object",
        "properties": {
          "dayOfWeek": {
            "type": "string",
            "enum": [
              "MONDAY",
              "TUESDAY",
              "WEDNESDAY",
              "THURSDAY",
              "FRIDAY",
              "SATURDAY",
              "SUNDAY"
            ],
            "example": "MONDAY"
          },
          "startTime": {
            "type": "string",
            "example": "10:00"
          },
          "endTime": {
            "type": "string",
            "example": "12:00"
          },
          "id": {
            "type": "string",
            "example": "c846...uuid"
          }
        },
        "required": [
          "dayOfWeek",
          "startTime",
          "endTime",
          "id"
        ]
      },
      "GetScheduleListResponseDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "멘토가 등록한 정기 스케줄 목록을 반환합니다."
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ScheduleItemResponseDto"
            }
          }
        },
        "required": [
          "message",
          "data"
        ]
      },
      "MentorReservationListResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "예약 ID",
            "example": "b2a9c1ef-27f4-4c02-ae34-0fcecc30c9f1"
          },
          "title": {
            "type": "string",
            "description": "세션 제목",
            "example": "React 입문 강의"
          },
          "date": {
            "type": "string",
            "description": "예약날짜",
            "example": "2025-07-27"
          },
          "startTime": {
            "type": "string",
            "description": "시작 시간",
            "example": "14:00:00"
          },
          "endTime": {
            "type": "string",
            "description": "종료 시간",
            "example": "15:00:00"
          },
          "status": {
            "type": "string",
            "description": "예약 상태",
            "example": "pending",
            "enum": [
              "pending",
              "confirmed",
              "rejected"
            ]
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "예약 생성 시간",
            "example": "2025-07-27T14:32:00.000Z"
          },
          "menteeName": {
            "type": "string",
            "description": "멘티 이름",
            "example": "김철수"
          },
          "menteeEmail": {
            "type": "string",
            "description": "멘티 이메일",
            "example": "mentee@example.com"
          },
          "menteePhone": {
            "type": "string",
            "description": "멘티 전화번호",
            "example": "010-1234-5678"
          }
        },
        "required": [
          "id",
          "title",
          "date",
          "startTime",
          "endTime",
          "status",
          "createdAt",
          "menteeName",
          "menteeEmail",
          "menteePhone"
        ]
      },
      "MentorReservationDetailResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "예약 ID",
            "example": "b2a9c1ef-27f4-4c02-ae34-0fcecc30c9f1"
          },
          "title": {
            "type": "string",
            "description": "세션 제목",
            "example": "React 입문 강의"
          },
          "date": {
            "type": "string",
            "description": "예약 날짜",
            "example": "2025-07-30"
          },
          "startTime": {
            "type": "string",
            "description": "시작 시간",
            "example": "14:00:00"
          },
          "endTime": {
            "type": "string",
            "description": "종료 시간",
            "example": "15:00:00"
          },
          "question": {
            "type": "string",
            "description": "멘티가 남긴 사전 질문",
            "example": "포트폴리오에 뭐 넣어야 할까요?"
          },
          "rejectReason": {
            "type": "string",
            "description": "거절 사유",
            "example": "요청 시간에 일정이 겹칩니다."
          },
          "status": {
            "type": "string",
            "description": "예약 상태",
            "example": "pending",
            "enum": [
              "pending",
              "confirmed",
              "rejected"
            ]
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "예약 생성 시간",
            "example": "2025-07-27T14:32:00.000Z"
          },
          "menteeName": {
            "type": "string",
            "description": "멘티 이름",
            "example": "김철수"
          },
          "menteeEmail": {
            "type": "string",
            "description": "멘티 이메일",
            "example": "mentee@example.com"
          },
          "menteePhone": {
            "type": "string",
            "description": "멘티 전화번호",
            "example": "010-1234-5678"
          }
        },
        "required": [
          "id",
          "title",
          "date",
          "startTime",
          "endTime",
          "question",
          "status",
          "createdAt",
          "menteeName",
          "menteeEmail",
          "menteePhone"
        ]
      },
...
      "CreateReviewDto": {
        "type": "object",
        "properties": {
          "reservationId": {
            "type": "string",
            "description": "예약 ID",
            "example": "uuid-string-here"
          },
          "rating": {
            "type": "number",
            "description": "평점 (1-5)",
            "example": 4,
            "minimum": 1,
            "maximum": 5
          },
          "content": {
            "type": "string",
            "description": "후기 내용",
            "example": "정말 좋았습니다."
          }
        },
        "required": [
          "reservationId",
          "rating",
          "content"
        ]
      },
      "UpdateReviewDto": {
        "type": "object",
        "properties": {}
      },
      "ReviewMyListItemDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "review-uuid"
          },
          "content": {
            "type": "string",
            "example": "정말 좋았습니다!"
          },
          "rating": {
            "type": "number",
            "example": 5
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "example": "2025-07-07T10:00:00Z"
          },
          "sessionTitle": {
            "type": "string",
            "example": "리액트강의"
          },
          "mentorName": {
            "type": "string",
            "example": "멘토닉네임"
          }
        },
        "required": [
          "id",
          "content",
          "rating",
          "createdAt",
          "sessionTitle",
          "mentorName"
        ]
      },
      "ReviewReceivedListItemDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "review-uuid"
          },
          "content": {
            "type": "string",
            "example": "정말 좋았습니다!"
          },
          "rating": {
            "type": "number",
            "example": 5
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "example": "2025-07-07T10:00:00Z"
          },
          "sessionTitle": {
            "type": "string",
            "example": "리액트 강의"
          },
          "menteeName": {
            "type": "string",
            "example": "멘티닉네임"
          }
        },
        "required": [
          "id",
          "content",
          "rating",
          "createdAt",
          "sessionTitle",
          "menteeName"
        ]
      },
...
      "ReviewListItemDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "review-uuid"
          },
          "content": {
            "type": "string",
            "example": "정말 좋았습니다!"
          },
          "rating": {
            "type": "number",
            "example": 5
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "example": "2025-07-07T10:00:00Z"
          }
        },
        "required": [
          "id",
          "content",
          "rating",
          "createdAt"
        ]
      },
...
    }
  }
}